Changes to eliminate any potential problems on 64-bit platforms.
2001-05-16  Mark Murnane  <Mark.Murnane@ireland.sun.com>

	* Changes to eliminate any potential problems on 64-bit platforms.

	* gnome-config.c (gnome_config_assemble_vector): Modified type
	of len to match return type of strlen().

	* gnome-ditem.c (replace_percentsign): Modified type of start,
	string_len and ps_len as they are used in strlen() operations.
	(strip_the_amp): Ditto for exec_len.
	(stripstreq): Ditto for len2.

	* gnome-exec.c (gnome_execute_async_with_env_fds): Modiifed type
	of res to gssize to match return type of read() and write() system
	calls.

	* gnome-moz-remote.c (mozilla_remote_test_window): Cast
	hostname to const char* to match g_strcasecmp's signature.

	* gnome-paper.c	(paper_name_compare): Modified return type
	from int to long to match return type of g_strcasecmp.
	(unit_name_compare): Ditto
	(unit_abbrev_compare): Ditto

	* gnome-program.c (gnome_program_install_property): Cast
	parameter #3 of both calls to g_param_spec_set_qdata.

	* gnome-score.c (log_score): Modified type of counter i to
	match return from strlen().

	* gnome-selector.c (_gnome_selector_add_history): Cast
	parameter #3 to GCompareFunc.

+++ libgnome-desktop/gnome-desktop-item.c
@@ -866,7 +866,7 @@ replace_percentsign(int argc, char **argv, const char *ps,
 		char *arg = argv[i];
 		char *p = strstr(arg,ps);
 		char *s;
-		int start, string_len, ps_len;
+		size_t start, string_len, ps_len;
 		if(!p) continue;
 		string_len = strlen(string);
 		ps_len = strlen(ps);
@@ -980,7 +980,7 @@ ditem_execute(const GnomeDesktopItem *item, int appargc, const char *appargv[], 
 static gboolean
 strip_the_amp(char *exec)
 {
-	int exec_len;
+	size_t exec_len;
 
 	g_strstrip(exec);
 	if(!*exec) return FALSE;
@@ -1134,7 +1134,7 @@ perc_replaced:
 static gboolean
 stripstreq(const char *s1, const char *s2)
 {
-	int len2;
+	size_t len2;
 
 	/* skip over initial spaces */
 	while(*s1 == ' ' || *s1 == '\t')
